Moved connectApplid and logonCics to ICicsTerminal
